tests and comments
